refactor(Makefiles): remove unneeded .PHONY target
