+++ Makefile
@@ -135,5 +135,3 @@ test-smoke:
 uninstall: check-fleet stop
 	$(FLEETCTL) unload -block-attempts=600 $(call deis_units,launched,.)
 	$(FLEETCTL) destroy $(strip $(call deis_units,.,.))
-
-.PHONY:
