+++ controller/Makefile
@@ -53,5 +53,3 @@ test-unit:
 
 test-functional:
 	GOPATH=$(CURDIR)/../tests/_vendor:$(GOPATH) go test -v ./tests/...
-
-.PHONY: test
