+++ router/tests/routerComponent_test.go
@@ -1,4 +1,4 @@
-package verbose
+package tests
 
 import (
 	"fmt"
